restructure the cronner codebase slightly
This moves the argument parsing out to its own file (`args.go`) to reduce the amount of code in `cronner.go`.
